+++ tests/testthat/test_miscUtils.R
@@ -1,0 +1,74 @@
+context("Tests for miscellaneous utilities in FSA")
+
+
+test_that("capFirst() capitalizations are correct",{
+  ## simulate data set
+  set.seed(345234534)
+  dbt <- data.frame(species=factor(rep(c("bluefin tuna"),30)),tl=round(rnorm(30,1900,300),0))
+  dbt$wt <- round(4.5e-05*dbt$tl^2.8+rnorm(30,0,6000),1)
+  dbg <- data.frame(species=factor(rep(c("Bluegill"),30)),tl=round(rnorm(30,130,50),0))
+  dbg$wt <- round(4.23e-06*dbg$tl^3.316+rnorm(30,0,10),1)
+  dlb <- data.frame(species=factor(rep(c("LMB"),30)),tl=round(rnorm(30,350,60),0))
+  dlb$wt <- round(2.96e-06*dlb$tl^3.273+rnorm(30,0,60),1)
+  df <- rbind(dbt,dbg,dlb)
+  df$rnd <- runif(nrow(df))
+  df$junk <- sample(c("Derek","Hugh","Ogle"),nrow(df),replace=TRUE)
+  ## actual tests
+  expect_equivalent(levels(factor(capFirst(df$species))),c("Bluefin Tuna","Bluegill","Lmb"))
+  expect_equivalent(levels(factor(capFirst(df$species,which="first"))),c("Bluefin tuna","Bluegill","Lmb"))
+})
+
+test_that("capFirst() returned classes are correct",{
+  ## simulate vector of names
+  vec <- c("Derek Ogle","derek ogle","Derek ogle","derek Ogle","DEREK OGLE")
+  fvec <- factor(vec)
+  ## first example of non-factor vector
+  vec1 <- capFirst(vec)
+  expect_equivalent(class(vec),class(vec1))
+  expect_equivalent(class(vec1),"character")
+  ## second example of factored vector
+  fvec1 <- capFirst(fvec)
+  expect_equivalent(class(fvec),class(fvec1))
+  expect_equivalent(class(fvec1),"factor")
+})
+
+test_that("lagratio() error messages",{
+  ## check error messages
+  expect_error(lagratio(0:5),"zeroes")
+  expect_error(lagratio(.leap.seconds),"POSIXt")
+  expect_error(lagratio(1:5,direction="derek"),"one of")
+  expect_error(lagratio(1:5,recursion=-1),"recursion")
+})
+
+test_that("lagratio() calculations",{
+  ## check calculations where latter is divided by former, no recursion
+  res1 <- (2:10)/(1:9)
+  res2 <- (3:10)/(1:8)
+  res3 <- (4:10)/(1:7)
+  expect_equal(lagratio(1:10,1),res1)
+  expect_equal(lagratio(1:10,2),res2)
+  expect_equal(lagratio(1:10,3),res3)
+  ## check calculations where latter is divided by former, with one level of recursion
+  res1r <- res1[2:length(res1)]/res1[1:(length(res1)-1)]
+  res2r <- res2[3:length(res2)]/res2[1:(length(res2)-2)]
+  res3r <- res3[4:length(res3)]/res3[1:(length(res3)-3)]
+  expect_equal(lagratio(1:10,1,2),res1r)
+  expect_equal(lagratio(1:10,2,2),res2r)
+  expect_equal(lagratio(1:10,3,2),res3r)
+  
+  ## check calculations where former is divided by latter, no recursion
+  res1 <- (1:9)/(2:10)
+  res2 <- (1:8)/(3:10)
+  res3 <- (1:7)/(4:10)
+  expect_equal(lagratio(1:10,1,direction="forward"),res1)
+  expect_equal(lagratio(1:10,2,direction="forward"),res2)
+  expect_equal(lagratio(1:10,3,direction="forward"),res3)
+  ## check calculations where latter is divided by former, with one level of recursion
+  res1r <- res1[1:(length(res1)-1)]/res1[2:length(res1)]
+  res2r <- res2[1:(length(res2)-2)]/res2[3:length(res2)]
+  res3r <- res3[1:(length(res3)-3)]/res3[4:length(res3)]
+  expect_equal(lagratio(1:10,1,2,direction="forward"),res1r)
+  expect_equal(lagratio(1:10,2,2,direction="forward"),res2r)
+  expect_equal(lagratio(1:10,3,2,direction="forward"),res3r)
+})
+